% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressTarXz.R
\name{compress.tar.xz}
\alias{compress.tar.xz}
\title{Compress a File or Directory Using the System's Own .tar.xz
  Installation}
\usage{
compress.tar.xz(sourcePath, targetPath = NULL, logging = TRUE)
}
\arguments{
\item{sourcePath}{the path to the file or folder to compress}

\item{targetPath}{the path to the archive to create}

\item{logging}{the logger to use}
}
\value{
the path to the newly created archive, or \code{NULL} on failure
}
\description{
Use the \code{tar} command installation of the operating system
  to compress a file or folder to a tar.xz archive of minimal size.
}
