% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileExtensionRegExp.R
\name{fileExtensionRegExp}
\alias{fileExtensionRegExp}
\title{Create a Regular Expression that can match a File Name versus One or
  Multiple File Extension(s)}
\usage{
fileExtensionRegExp(extensions)
}
\arguments{
\item{extensions}{either a single string or array of strings, each holding a
file extension}
}
\value{
the regular expression which will match them
}
\description{
For a given list \code{extensions} of file extensions, e.g.,
  \code{c("txt", "csv")} or just a single file extension, this function
  creates a regular expression while will match a file path that ends in any
  of the extensions (with a leading ".").
}
\examples{
exp <- fileExtensionRegExp(c("html", "jpg"))
exp
# [1] "^.*\\\\.(html|jpg)$"
length(grep(exp, "hellojpg")) > 0L
# [1] FALSE
length(grep(exp, "hello.jpg")) > 0L
# [1] TRUE
length(grep(exp, "hello.html")) > 0L
# [1] TRUE
length(grep(exp, "hellohtml")) > 0L
# [1] FALSE
length(grep(exp, "hello.htm")) > 0L
# [1] FALSE
length(grep(exp, "hello.html.txt")) > 0L
# [1] FALSE
}
