% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findFinite.R
\name{find.finite.nw}
\alias{find.finite.nw}
\title{Try to Find a Finite Value for a function \code{f}, and suppress
  warnings}
\usage{
find.finite.nw(x1, x2, f)
}
\arguments{
\item{x1}{the \code{x} coordinate at which we want to evaluate \code{f} (and
also the first interval limit)}

\item{x2}{the second interval limit, should be either larger or smaller than
\code{x1}}

\item{f}{the unary function}
}
\value{
a tuple \code{x, f(x)}, which results from the attempt to find a
  finite function value
}
\description{
Given is a unary function \code{f} and a coordinate \code{x1} as
  well as a value \code{x2}. The general goal is to find an \code{x} value
  for which \code{f} becomes finite and which is as close as possible to
  \code{x1} from the range \code{min(x1,x2)...max(x1,x2)}. If \code{f(x1)} is
  finite, return \code{c(x1, f(x1)}. Otherwise, try to find a value \code{t}
  within the interval delimited by \code{x1} and \code{x2} for which \code{f}
  takes on a finite value, then return \code{c(t, f(t)}. If no such value can
  be discovered, return \code{c(x1, f(x1)}. This function will produce no
  warnings, at the cost of slower performance (see
  \code{\link{find.finite}}).
}
\seealso{
find.finite
}
