% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionArgs.R
\name{function.args}
\alias{function.args}
\title{Get the Names of the Arguments of a Function}
\usage{
function.args(f)
}
\arguments{
\item{f}{the function}
}
\value{
the vector of function arguments, or \code{NULL} if the function has
  no arguments
}
\description{
Get the names of the arguments of a function as a vector of
  strings.
}
\examples{
function.args(sin)
# [1] "x"
f <- function(x) sin(x);
function.args(f)
# [1] "x"
f <- function() { }
function.args(f)
# NULL
}
