% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionName.R
\name{function.name}
\alias{function.name}
\title{Get the Name of a Function as String}
\usage{
function.name(f)
}
\arguments{
\item{f}{the function}
}
\value{
a string representing the name of the function
}
\description{
Get the Name of a Function as String
}
\examples{
function.name(sin)
# sin(x)
library(methods);
function.name(setClass)
# methods::setClass
f <- function(k) sin(k);
ff <- f;
function.name(ff);
# f
nest <- function(z) function.name(z);
nest(ff);
# f
nest(sin);
# sin
}
