% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionToString.R
\name{function.toString}
\alias{function.toString}
\title{Convert a Function Body to a String}
\usage{
function.toString(f)
}
\arguments{
\item{f}{the function}
}
\value{
a string representation of the function
}
\description{
Convert a function body to a string
}
\examples{
function.toString(function(x,y) x+5*y)
# x+5*y
function.toString(function(x) sin(x))
# sin(x)
function.toString(sin)
# sin
}
