% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{functionOrNULL-class}
\alias{functionOrNULL-class}
\title{Either a \code{function} or \code{NULL}}
\description{
A combination of the \code{function} class and \code{NULL}.
  Using this class allows us to have member variables in S4 classes which are
  null-able functions.
}
