% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionToString.R
\name{functionToString}
\alias{functionToString}
\title{Convert a Function Body to a String}
\usage{
functionToString(f)
}
\arguments{
\item{f}{the function}
}
\value{
a string representation of the function
}
\description{
Convert a function body to a string
}
