% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignoreErrors.R
\name{ignoreErrors}
\alias{ignoreErrors}
\title{Execute the Expression and Ignore All Errors}
\usage{
ignoreErrors(exp)
}
\arguments{
\item{exp}{the expression}
}
\description{
Execute an expression and ignore all errors.
}
