% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is.non.empty.list}
\alias{is.non.empty.list}
\title{Check whether Something is a non-empty list}
\usage{
is.non.empty.list(t)
}
\arguments{
\item{t}{the object to check}
}
\value{
\code{TRUE} if \code{is.null(t)} does not hold
  and \code{is.list(t)} and \code{length(t) > 0L} hold, \code{FALSE}
  otherwise
}
\description{
Check whether an object is a non-empty list.
}
