% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is.non.empty.string}
\alias{is.non.empty.string}
\title{Check whether Something is a non-empty string}
\usage{
is.non.empty.string(t)
}
\arguments{
\item{t}{the object to check}
}
\value{
\code{TRUE} if neither \code{is.na(t)} nor \code{is.null(t)} holds
  and \code{is.character(t)} and \code{length(t) > 0L} hold, \code{FALSE}
  otherwise
}
\description{
Check whether an object is a non-empty string.
}
