% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is.non.empty.vector}
\alias{is.non.empty.vector}
\title{Check whether Something is a non-empty vector}
\usage{
is.non.empty.vector(t)
}
\arguments{
\item{t}{the object to check}
}
\value{
\code{TRUE} if neither \code{is.na(t)} nor \code{is.null(t)} holds
  and \code{is.vector(t)} and \code{length(t) > 0L} hold, \code{FALSE}
  otherwise
}
\description{
Check whether an object is a non-empty vector.
}
