% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is.not.na.or.null}
\alias{is.not.na.or.null}
\title{Check whether Something is NOT \code{NULL} or \code{NA}}
\usage{
is.not.na.or.null(t)
}
\arguments{
\item{t}{the object to check}
}
\value{
\code{TRUE} if neither \code{is.na(t)} nor \code{is.null(t)} holds,
  \code{FALSE} otherwise
}
\description{
Check whether an object is NOT \code{NULL} or \code{NA}.
}
