% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLogger.R
\name{makeLogger}
\alias{makeLogger}
\title{Create a Logger Function}
\usage{
makeLogger(logging, cores = 1L)
}
\arguments{
\item{logging}{should progress information be printed: either \code{TRUE} for
printing to the console via \code{\link{print}}, \code{FALSE} for no
logging, or a path to a file receiving logging information}

\item{cores}{the number of processing cores. This is not really relevant
itself, but (\code{>1L} will issue a warning for console-based logging.}
}
\description{
Create a logger function which can be used to store status
  output either to a file or stdout.
}
