% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{numericOrNULL-class}
\alias{numericOrNULL-class}
\title{Either a \code{numeric} \code{vector} or \code{NULL}}
\description{
A combination of the \code{numeric} class and \code{NULL}. Using
  this class allows us to have member variables in S4 classes which are
  null-able numeric vectors.
}
