% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchProcessor.R
\name{path.batchProcessor}
\alias{path.batchProcessor}
\title{Create a Batch Processor}
\usage{
path.batchProcessor(processor, dest, suffix)
}
\arguments{
\item{processor}{the \code{function(path(s), dest)} to be applied}

\item{dest}{the destination folder}

\item{suffix}{the suffix}
}
\value{
the batch processor function of type \code{function(path(s), dest)}
}
\description{
Create a batch processor function for
  \code{\link{path.batchApply}} by mapping a function \code{processor} of
  type \code{function(path(s), dest)} to a function \code{function(root,
  path(s))} for a given basic destination directory \code{dest} and adding a
  \code{suffix} to a given set of paths.
}
