% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativize.R
\name{path.relativize}
\alias{path.relativize}
\title{Relativize a Path versus another Path}
\usage{
path.relativize(path, dir = getwd(), sep = .Platform$file.sep)
}
\arguments{
\item{path}{the path to be normalized}

\item{dir}{the reference directory to normalize against}

\item{sep}{path separator}
}
\description{
Relativizes a path \code{path} versus a directory \code{dir}. Both
\code{path} and \code{dir} are normalized. If \code{dir} and \code{path}
identify the same object, \code{""} is returned. Otherwise, if the returned
path is appended to \code{dir} via \code{\link{file.path}}, the resulting
path should identify \code{path}.
}
\details{
Inspired by the function of the same name in the \code{nonmemica} package,
see https://github.com/cran/nonmemica/blob/master/R/path.R.
}
