% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonPrefix.R
\name{string.commonPrefix}
\alias{string.commonPrefix}
\title{Find the Longest Common Prefix of a Set of Strings}
\usage{
string.commonPrefix(strings)
}
\arguments{
\item{strings}{the strings}
}
\value{
the longest common prefix string
}
\description{
For a set \code{strings} of strings, this method returns the
  longest common prefix. Inspired by
  \url{http://stackoverflow.com/questions/27470481/} and
  \url{http://stackoverflow.com/questions/26285010}
}
